Abstract the zoomable / pannable / scrollable timeline view in a separate superclass
