+++ static/views/GanttView/GanttView.js
@@ -1,15 +1,9 @@
-/* globals uki, d3 */
-import LinkedMixin from '../common/LinkedMixin.js';
-import CursoredViewMixin from '../common/CursoredViewMixin.js';
-import normalizeWheel from '../../utils/normalize-wheel.js';
-import cleanupAxis from '../../utils/cleanupAxis.js';
+/* globals d3 */
+import ZoomableTimelineView from '../ZoomableTimelineView/ZoomableTimelineView.js';
 
 // Minimum vertical pixels per row
 const MIN_LOCATION_HEIGHT = 30;
 
-// Fetch and draw 3x the time data than we're actually showing, for smooth
-// scrolling, zooming interactions
-const HORIZONTAL_SPILLOVER_FACTOR = 3;
 // Fetch and draw 3x the time data than we're actually showing, for smooth
 // scrolling interactions
 const VERTICAL_SPILLOVER_FACTOR = 3;
@@ -17,33 +11,15 @@ const VERTICAL_SPILLOVER_FACTOR = 3;
 // Don't show trace lines when we're zoomed out beyond this time limit
 const TRACE_LINE_TIME_LIMIT = Infinity;
 
-class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated through app-wide things like Controller.refreshDatasets()
-  CursoredViewMixin( // Adds and updates a line in the background wherever the user is mousing
-    uki.ui.ParentSizeViewMixin( // Keeps the SVG element sized based on how much space GoldenLayout gives us
-      uki.ui.SvgGLView))) { // Ensures this.d3el is an SVG element; adds the download icon to the tab
+class GanttView extends ZoomableTimelineView { // abstracts a lot of common logic for smooth zooming + panning + rendering offscreen + showing scrollbars for timeline-based views
   constructor (options) {
     options.resources = (options.resources || []).concat(...[
-      { type: 'less', url: 'views/GanttView/style.less' },
-      { type: 'text', url: 'views/GanttView/template.svg', name: 'template' },
       // Placeholder resources that don't actually get updated until later
       { type: 'placeholder', value: null, name: 'totalUtilization' },
       { type: 'placeholder', value: null, name: 'selectionUtilization' },
       { type: 'placeholder', value: null, name: 'selectedIntervalTrace' }
     ]);
     super(options);
-
-    // Ensure unique clip path IDs for each GanttView instantiation (can create
-    // problems if there's more than one GanttView)
-    this.clipPathId = (GanttView.NEXT_CLIP_ID || 1);
-    GanttView.NEXT_CLIP_ID += 1;
-    this.clipPathId = 'clip' + this.clipPathId;
-
-    this.margin = {
-      top: 20,
-      right: 20,
-      bottom: 40,
-      left: 40
-    };
 
     // yScale maps the full list of locationNames to the full height of the
     // canvas
@@ -71,14 +47,6 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
       }
       return result;
     };
-
-    // xScale refers to the data that's visible; converts from timestamps to
-    // the width of .yScroller
-    this.xScale = d3.scaleLinear();
-    // xFakeScrollerScale refers to the full range of the data; converts from
-    // timestamps to how wide the canvas *would* be if we could fit it all
-    // onscreen (used only for the horizontal scrollbar)
-    this.xFakeScrollerScale = d3.scaleLinear();
   }
 
   get isLoading () {
@@ -116,272 +84,103 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
     }
   }
 
-  async setup () {
-    await super.setup(...arguments);
+  handlePanningStart (event, dragState) {
+    dragState.y0 = event.y;
+    dragState.dy = 0;
+    dragState.initialYScroll = this.d3el.select('foreignObject').node().scrollTop;
+  }
 
-    // Set up the SVG element and position its .chart group
-    this.d3el.html(this.getNamedResource('template'))
-      .classed('GanttView', true);
-    this.d3el.select('.chart')
-      .attr('transform', `translate(${this.margin.left},${this.margin.top})`);
+  handlePanning (event, dragState, newDomain) {
+    dragState.dy = event.y - this._dragState.y0;
+    const scrollTop = dragState.initialYScroll - dragState.dy;
+    const forceQuickDraw = scrollTop !== this.d3el.select('foreignObject').node().scrollTop;
+    this._ignoreYScrollEvents = true;
+    this.d3el.select('foreignObject').node().scrollTop = scrollTop;
+    if (forceQuickDraw &&
+        newDomain[0] === this.linkedState.detailDomain[0] &&
+        newDomain[1] === this.linkedState.detailDomain[1]) {
+      // TracedLinkedState won't otherwise issue a quickDraw in this case,
+      // which can result in some funny effects if there was vertical
+      // panning
+      this.quickDraw();
+    }
+  }
 
-    // Link our y axis and the clip region
-    this.d3el.select('clipPath').attr('id', this.clipPathId);
-    this.d3el.select('.yAxis').attr('clip-path', `url(#${this.clipPathId})`);
-
-    // Insert a div inside the goldenlayout wrapper to put a scrollbar
-    // *below* the x-axis
-    this.xFakeScroller = this.glEl.append('div')
-      .classed('xFakeScroller', true)
-      .style('left', this.margin.left + 'px');
-    // Empty div inside to make CSS overflow-x work
-    this.xFakeScroller.append('div')
-      .classed('scrollContent', true);
-
-    // Convenience pointer to .yScroller
-    this.yScroller = this.d3el.select('.yScroller');
-
-    // Do a quickDraw immediately for horizontal brush / scroll / zoom
-    // interactions...
-    this.linkedState.on('detailDomainChangedSync', () => { this.quickDraw(); });
-    // ... and ask for new data when we're confident that rapid interactions
-    // have finished
-    this.linkedState.on('detailDomainChanged', () => {
-      this.updateDataIfNeeded();
-    });
-    // Also ask for new data when the selection changes
-    this.linkedState.on('selectionChanged', () => {
-      this.updateDataIfNeeded();
-    });
-
-    // Set up local zoom, pan, hover, and click interactions
-    this.setupInteractions();
-
-    // Set up the cursor
-    this.setupCursor(this.d3el.select('.chart'));
-
-    // setup() is only called once this.d3el is ready; only at this point do we
-    // know how many bins to ask for
-    this.updateDataIfNeeded();
+  handlePanningEnd (event, dragState) {
+    if (dragState.dx === 0 && dragState.dy === 0) {
+      const timestamp = Math.round(this.xScale.invert(event.x));
+      const location = this.yScale.invert(event.y + this.d3el.select('foreignObject').node().scrollTop);
+      this.linkedState.selectInterval(timestamp, location);
+    }
   }
 
   setupInteractions () {
-    // Update whenever GoldenLayout resizes us
-    this.on('resize', () => { this.render(); });
-
-    this.d3el.select('.eventCapturer')
-      .on('wheel', event => {
-        // Zoom when using the wheel over the main chart area
-        const zoomFactor = 1.05 ** (normalizeWheel(event).pixelY / 100);
-        // Where was the mouse, relative to the chart (not its actual target;
-        // .eventCapturer wouldn't work)
-        const chartBounds = this.d3el.select('.chart').node().getBoundingClientRect();
-        const mousedPosition = this.xScale.invert(event.clientX - chartBounds.left);
-        this.linkedState.detailDomain = [
-          mousedPosition - zoomFactor * (mousedPosition - this.linkedState.detailDomain[0]),
-          mousedPosition + zoomFactor * (this.linkedState.detailDomain[1] - mousedPosition)
-        ];
-      }, { passive: true }).call(d3.drag()
-        .on('start', event => {
-          const initialDomain = this.linkedState.detailDomain;
-          const initialTimespan = initialDomain[1] - initialDomain[0];
-          this._dragState = {
-            initialDomain,
-            initialTimespan,
-            initialYScroll: this.yScroller.node().scrollTop,
-            x0: event.x,
-            y0: event.y,
-            dx: 0,
-            dy: 0
-          };
-          // Prevent false scroll events from changing scrollTop as a result of
-          // dragging
-          this._ignoreYScrollEvents = true;
-        }).on('drag', event => {
-          // Vertical dragging
-          this._dragState.dy = event.y - this._dragState.y0;
-          const scrollTop = this._dragState.initialYScroll - this._dragState.dy;
-          const forceQuickDraw = scrollTop !== this.yScroller.node().scrollTop;
-          this.yScroller.node().scrollTop = scrollTop;
-          // Horizontal dragging
-          this._dragState.dx = event.x - this._dragState.x0;
-          const mouseDelta = this.xScale.invert(event.x) -
-            this.xScale.invert(event.x + this._dragState.dx);
-          const newDomain = [
-            this._dragState.initialDomain[0] + mouseDelta,
-            this._dragState.initialDomain[1] + mouseDelta
-          ];
-          // Prevent zooming when dragging to the end of the screen
-          if (newDomain[0] <= this.linkedState.overviewDomain[0]) {
-            newDomain[0] = this.linkedState.overviewDomain[0];
-            newDomain[1] = this.linkedState.overviewDomain[0] +
-              this._dragState.initialTimespan;
-          } else if (newDomain[1] >= this.linkedState.overviewDomain[1]) {
-            newDomain[1] = this.linkedState.overviewDomain[1];
-            newDomain[0] = this.linkedState.overviewDomain[1] -
-              this._dragState.initialTimespan;
-          }
-          if (forceQuickDraw &&
-              newDomain[0] === this.linkedState.detailDomain[0] &&
-              newDomain[1] === this.linkedState.detailDomain[1]) {
-            // TracedLinkedState won't otherwise issue a quickDraw in this case,
-            // which can result in some funny effects if there was vertical
-            // panning
-            this.quickDraw();
-          }
-          this.linkedState.detailDomain = newDomain;
-        }).on('end', event => {
-          if (this._dragState.dx === 0 && this._dragState.dy === 0) {
-            const timestamp = Math.round(this.xScale.invert(event.x));
-            const location = this.yScale.invert(event.y + this.yScroller.node().scrollTop);
-            this.linkedState.selectInterval(timestamp, location);
-          }
-          delete this._dragState;
-          delete this._ignoreYScrollEvents;
-          this.render();
-        }));
-
-    // Pan the detailDomain in response to scrolling the x axis
-    this.xFakeScroller.on('scroll', () => {
-      if (this._ignoreXScrollEvents) {
-        // This was an artificial scroll event from setting element.scrollLeft;
-        // ignore it
-        this._ignoreXScrollEvents = false;
-        return;
-      }
-      const scrollLeft = this.xFakeScroller.node().scrollLeft;
-      // Update the domain based on where we've scrolled to
-      const oldDomain = this.linkedState.detailDomain;
-      const left = this.xFakeScrollerScale.invert(scrollLeft);
-      this.linkedState.detailDomain = [left, left + (oldDomain[1] - oldDomain[0])];
-    });
+    super.setupInteractions();
 
     // Make sure the y axis links with scrolling
-    this.yScroller.on('scroll', () => {
-      if (!this._ignoreYScrollEvents) {
-        this.quickDraw();
-        this.render();
+    this.d3el.select('foreignObject').on('scroll', () => {
+      if (this._ignoreYScrollEvents) {
+        // suppress false scroll events from setting scrollTop
+        this._ignoreYScrollEvents = false;
+        return;
       }
+      this.quickDraw();
+      this.render();
     });
     // Link wheel events on the y axis back to vertical scrolling
     this.d3el.select('.yAxisScrollCapturer').on('wheel', event => {
-      this.yScroller.node().scrollTop += event.deltaY;
+      this.d3el.select('foreignObject').node().scrollTop += event.deltaY;
     });
   }
 
-  /**
-   * This is called immediately for rapid updates during things like zooming or
-   * scrolling (so this should never include expensive drawing commands)
-   */
-  quickDraw () {
-    const chartShape = this.getChartShape();
-    this.moveCanvas(chartShape);
-    this.drawAxes(chartShape);
+  drawCanvas (chartShape) {
+    this.drawTraceLines(chartShape);
+    this.drawBars(chartShape);
   }
 
-  async draw () {
-    await super.draw(...arguments);
-
-    if (this.isLoading || this.error) {
-      // Don't draw anything if we're still waiting on something; super.draw
-      // will show a spinner. Or if there's an upstream error, super.draw will
-      // already display an error message. Don't attempt to draw anything (or
-      // we'll probably just add to the noise of whatever is really wrong)
-      return;
-    }
-
-    // Store the shape that was last fully rendered (note that this is different
-    // from _lastChartShape, which indicates which shape was last queried)
-    delete this._renderedChartShape;
-    this._renderedChartShape = this.getChartShape();
-
-    this.drawCursor();
-    this.moveCanvas(this._renderedChartShape);
-    this.drawAxes(this._renderedChartShape);
-    // More expensive rendering that should only happen here, not quickDraw
-    this.prepCanvas(this._renderedChartShape);
-    this.drawTraceLines(this._renderedChartShape);
-    this.drawBars(this._renderedChartShape);
-
-    // As part of the full render (that could have been triggered from anywhere,
-    // start the process of requesting fresh data if the viewport is out of date)
-    this.updateDataIfNeeded(this._renderedChartShape);
+  hasEnoughDataToComputeChartShape () {
+    return super.hasEnoughDataToComputeChartShape() &&
+      !!this.linkedState.info.locationNames;
   }
 
-  /**
-   * Checks to see if we need to request new data
-   */
-  async updateDataIfNeeded (chartShape) {
-    if (!this.linkedState.detailDomain || !this.linkedState.info.locationNames) {
-      // We don't have enough information to know what data to ask for (e.g. the
-      // server might still be bundling a large dataset); wait for Controller.js
-      // to refreshDatasets()
-      return;
-    }
-    if (!chartShape) {
-      chartShape = this.getChartShape();
-    }
+  determineIfShapeNeedsRefresh (lastChartShape, chartShape) {
+    return lastChartShape.locations.length !== chartShape.locations.length ||
+    lastChartShape.locations.some((loc, i) => chartShape.locations[i] !== loc);
+  }
+
+  async updateData (chartShape) {
     const domain = chartShape.spilloverXScale.domain();
-    let locations = chartShape.locations;
 
-    // We consider chartShape to be out of date if:
-    // 1. the spillover domain is different,
-    const needsRefresh = !this._lastChartShape ||
-      this._lastChartShape.spilloverXScale.domain()
-        .some((timestamp, i) => domain[i] !== timestamp) ||
-    // 2. the expected locations are different,
-      this._lastChartShape.locations.length !== locations.length ||
-      this._lastChartShape.locations.some((loc, i) => locations[i] !== loc) ||
-    // 3. the selection has changed
-      this._lastSelectionId !== this.linkedState.selection?.id;
+    // Make the list of locations a URL-friendly comma-separated list
+    const locations = globalThis.encodeURIComponent(chartShape.locations.join(','));
 
-    if (needsRefresh) {
-      // Cache the shape and selection ids that are currently relevant, so we
-      // know if something changed since last time
-      this._lastChartShape = chartShape;
-      this._lastSelectionId = this.linkedState.selection?.id;
+    // Basic URL that both totalUtilization and selectionUtilization will use
+    const baseUrl = `/datasets/${this.datasetId}/utilizationHistogram?bins=${chartShape.bins}&begin=${domain[0]}&end=${domain[1]}&locations=${locations}`;
 
-      // Make the list of locations a URL-friendly comma-separated list
-      locations = globalThis.encodeURIComponent(locations.join(','));
+    // Add any additional per-selection parameters
+    const selectionParams = this.linkedState.selection?.utilizationParameters;
 
-      // Basic URL that both totalUtilization and selectionUtilization will use
-      const baseUrl = `/datasets/${this.datasetId}/utilizationHistogram?bins=${chartShape.bins}&begin=${domain[0]}&end=${domain[1]}&locations=${locations}`;
+    // Send the utilization API requests
+    const totalPromise = this.updateResource({ name: 'totalUtilization', type: 'json', url: baseUrl });
+    const selectionPromise = selectionParams
+      ? this.updateResource({ name: 'selectionUtilization', type: 'json', url: baseUrl + selectionParams })
+      : this.updateResource({ name: 'selectionUtilization', type: 'placeholder', value: null }); // no current selection; replace data with null
 
-      // Add any additional per-selection parameters
-      const selectionParams = this.linkedState.selection?.utilizationParameters;
+    // Update the traceback data for the selected interval (if there is one)
+    const selectedIntervalId = this.linkedState.selection?.intervalDetails?.intervalId;
+    const tracebackPromise = selectedIntervalId
+      ? this.updateResource({
+          name: 'selectedIntervalTrace',
+          type: 'json',
+          url: `/datasets/${this.datasetId}/intervals/${selectedIntervalId}/trace?begin=${domain[0]}&end=${domain[1]}`
+        })
+      : this.updateResource({ name: 'selectedIntervalTrace', type: 'placeholder', value: null });
 
-      // Send the utilization API requests
-      const totalPromise = this.updateResource({ name: 'totalUtilization', type: 'json', url: baseUrl });
-      const selectionPromise = selectionParams
-        ? this.updateResource({ name: 'selectionUtilization', type: 'json', url: baseUrl + selectionParams })
-        : this.updateResource({ name: 'selectionUtilization', type: 'placeholder', value: null }); // no current selection; replace data with null
-
-      // Update the traceback data for the selected interval (if there is one)
-      const selectedIntervalId = this.linkedState.selection?.intervalDetails?.intervalId;
-      const tracebackPromise = selectedIntervalId
-        ? this.updateResource({
-            name: 'selectedIntervalTrace',
-            type: 'json',
-            url: `/datasets/${this.datasetId}/intervals/${selectedIntervalId}/trace?begin=${domain[0]}&end=${domain[1]}`
-          })
-        : this.updateResource({ name: 'selectedIntervalTrace', type: 'placeholder', value: null });
-
-      // Initial render call to show the spinner if waiting for data takes a while
-      // (because render() is debounced, the spinner won't show if the request
-      // is fast)
-      this.render();
-      await Promise.all([totalPromise, selectionPromise, tracebackPromise]);
-      // Ensure that everything is updated with the new data
-      this.render();
-    }
-    return chartShape;
+    return Promise.all([totalPromise, selectionPromise, tracebackPromise]);
   }
 
-  computeSpillover ([low, high], factor) {
-    const halfOriginalWidth = (high - low) / 2;
-    const center = low + halfOriginalWidth;
-    const halfSpilloverWidth = factor * halfOriginalWidth;
-    return [Math.floor(center - halfSpilloverWidth), Math.ceil(center + halfSpilloverWidth)];
+  getRequiredChartHeight () {
+    return MIN_LOCATION_HEIGHT * this.linkedState.info.locationNames.length;
   }
 
   /**
@@ -393,146 +192,27 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
    * the last updateShapeAndDataIfNeeded call)
    */
   getChartShape () {
-    // Figure out how much space we have, including whether or not to show the
-    // scrollbars
-    const bounds = this.getBounds();
-    const chartShape = {
-      chartWidth: bounds.width - this.margin.left - this.margin.right,
-      chartHeight: bounds.height - this.margin.top - this.margin.bottom,
-      requiredHeight: MIN_LOCATION_HEIGHT * this.linkedState.info.locationNames.length
-    };
-    chartShape.rightScrollbarIsShowing = chartShape.requiredHeight > chartShape.chartHeight;
-    chartShape.bottomScrollbarIsShowing =
-      this.linkedState.detailDomain[0] > this.linkedState.overviewDomain[0] ||
-      this.linkedState.detailDomain[1] < this.linkedState.overviewDomain[1];
-    if (chartShape.rightScrollbarIsShowing) {
-      chartShape.chartWidth -= this.scrollBarSize;
-    }
-    if (chartShape.bottomScrollbarIsShowing) {
-      chartShape.chartHeight -= this.scrollBarSize;
-    }
-    // Force at least 1 px width, height
-    chartShape.chartWidth = Math.max(1, chartShape.chartWidth);
-    chartShape.chartHeight = Math.max(1, chartShape.chartHeight);
-    // Use either the required space or expand to use the space we have
-    chartShape.fullHeight = Math.max(chartShape.requiredHeight, chartShape.chartHeight);
+    const chartShape = super.getChartShape();
 
-    // Update the scales
-    this.xScale.range([0, chartShape.chartWidth])
-      .domain(this.linkedState.detailDomain);
     this.yScale.range([0, chartShape.fullHeight])
       .domain(this.linkedState.info.locationNames);
 
-    // How many pixels would the full data span at this zoom level?
-    const overviewRange = this.linkedState.overviewDomain.map(this.xScale);
-    chartShape.fullWidth = overviewRange[1] - overviewRange[0];
-    chartShape.overviewScale = d3.scaleLinear()
-      .domain(this.linkedState.overviewDomain)
-      .range(overviewRange);
-    // Update this.xFakeScrollerScale while we're at it
-    this.xFakeScrollerScale
-      .domain(this.linkedState.overviewDomain)
-      .range([0, chartShape.fullWidth]);
-
-    // Figure out the data / pixel bounds that we should query
-    const spilloverXDomain = this.computeSpillover(this.linkedState.detailDomain, HORIZONTAL_SPILLOVER_FACTOR);
-    // Ensure integer endpoints
-    spilloverXDomain[0] = Math.floor(spilloverXDomain[0]);
-    spilloverXDomain[1] = Math.ceil(spilloverXDomain[1]);
-    const spilloverXRange = spilloverXDomain.map(chartShape.overviewScale);
-    chartShape.spilloverXScale = d3.scaleLinear()
-      .domain(spilloverXDomain)
-      .range(spilloverXRange);
-
-    // How many (integer) bins should we ask for (1 per pixel)?
-    chartShape.bins = Math.ceil(spilloverXRange[1] - spilloverXRange[0]);
-
     // Given the scroll position and size, which locations should be visible?
+    const scrollTop = this.d3el.select('foreignObject').node().scrollTop;
     let spilloverYRange = [
-      this.yScroller.node().scrollTop,
-      this.yScroller.node().scrollTop + chartShape.chartHeight
+      scrollTop,
+      scrollTop + chartShape.chartHeight
     ];
     // Add vertical spillover
     spilloverYRange = this.computeSpillover(spilloverYRange, VERTICAL_SPILLOVER_FACTOR);
     chartShape.spilloverYRange = spilloverYRange;
     chartShape.locations = this.yScale.invertRange(...spilloverYRange);
 
-    // Figure out what transformations we need relateive to the last time a
-    // full render happened
-    if (this._renderedChartShape) {
-      const originalDomain = this._renderedChartShape.spilloverXScale.domain();
-      chartShape.zoomFactor = (originalDomain[1] - originalDomain[0]) / (spilloverXDomain[1] - spilloverXDomain[0]);
-      chartShape.leftOffset = chartShape.spilloverXScale(this._renderedChartShape.spilloverXScale.domain()[0]);
-    } else {
-      chartShape.zoomFactor = 1.0;
-      chartShape.leftOffset = spilloverXRange[0];
-    }
-
     return chartShape;
   }
 
-  getMousedTime (offsetX) {
-    return this.xScale.invert(offsetX - this.margin.left);
-  }
-
-  getCursorHeight () {
-    return this._renderedChartShape?.chartHeight || 0;
-  }
-
-  getCursorPosition (time) {
-    return time < this.xScale.domain()[0] || time > this.xScale.domain()[1]
-      ? null
-      : this.xScale(time);
-  }
-
-  moveCanvas (chartShape) {
-    // Note that we resize the canvas using CSS transforms instead of width /
-    // height attributes to get a scaling effect while we're zooming / panning.
-    // CSS transforms already take advantage of the GPU; see also:
-    // https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API/Tutorial/Optimizing_canvas#scaling_canvas_using_css_transforms
-    this.d3el.select('.gantt-canvas')
-      .style('transform-origin', '0px 0px')
-      .style('transform', `scale(${chartShape.zoomFactor}, 1)`)
-      .style('left', `${chartShape.leftOffset}px`);
-
-    // Update the eventCapturer rect
-    this.d3el.select('.eventCapturer')
-      .attr('width', chartShape.chartWidth)
-      .attr('height', chartShape.chartHeight);
-
-    // Update the size of the foreignObject
-    this.yScroller
-      .attr('width', chartShape.chartWidth +
-        (chartShape.rightScrollbarIsShowing ? this.scrollBarSize : 0))
-      .attr('height', chartShape.chartHeight)
-      .style('overflow-y', chartShape.rightScrollbarIsShowing ? 'scroll' : 'hidden');
-    // Update the fake scroller position, size
-    this.xFakeScroller
-      .style('width', chartShape.chartWidth + 'px')
-      .style('top', (this.margin.top + chartShape.chartHeight) + 'px')
-      .style('bottom', '0px')
-      .style('overflow-x', chartShape.bottomScrollbarIsShowing ? 'scroll' : 'hidden');
-    // Update the empty div inside to simulate the correct scrollbar position
-    this.xFakeScroller.select('.scrollContent')
-      .style('width', chartShape.fullWidth + 'px')
-      .style('height', chartShape.fullHeight + 'px');
-    // Scroll the empty div to the current position
-    const scrollLeft = this.xFakeScrollerScale(this.linkedState.detailDomain[0]);
-    this._ignoreXScrollEvents = true;
-    this.xFakeScroller.node().scrollLeft = scrollLeft;
-  }
-
   drawAxes (chartShape) {
-    // Update the x axis
-    const xAxisGroup = this.d3el.select('.xAxis')
-      .attr('transform', `translate(0, ${chartShape.chartHeight})`)
-      .call(d3.axisBottom(this.xScale));
-    cleanupAxis(xAxisGroup);
-
-    // Position the x label
-    this.d3el.select('.xAxisLabel')
-      .attr('x', chartShape.chartWidth / 2)
-      .attr('y', chartShape.chartHeight + 2 * this.emSize);
+    super.drawAxes(chartShape);
 
     // Update the y axis
     let yTicks = this.d3el.select('.yAxis').selectAll('.tick')
@@ -541,23 +221,6 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
     const yTicksEnter = yTicks.enter().append('g')
       .classed('tick', true);
     yTicks = yTicks.merge(yTicksEnter);
-
-    // Link the y axis position to yScroller
-    const yOffset = this.yScroller.node().scrollTop;
-    this.d3el.select('.yAxis').attr('transform', `translate(0,${-yOffset})`);
-
-    // Update the clipPath position and dimensions
-    this.d3el.select('clipPath rect')
-      .attr('x', -this.margin.left)
-      .attr('y', yOffset)
-      .attr('width', chartShape.chartWidth + this.margin.left)
-      .attr('height', chartShape.chartHeight);
-
-    // Update .yAxisScrollCapturer
-    this.d3el.select('.yAxisScrollCapturer')
-      .attr('x', -this.margin.left)
-      .attr('width', this.margin.left)
-      .attr('height', chartShape.chartHeight);
 
     // y tick coordinate system in between each row
     yTicks.attr('transform', d => `translate(0,${this.yScale(d) + this.yScale.bandwidth() / 2})`);
@@ -587,20 +250,9 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
         return aggText;
       });
 
-    // Position the y label
+    // Set the y label
     this.d3el.select('.yAxisLabel')
-      .attr('transform', `translate(${-this.emSize - 12},${chartShape.chartHeight / 2}) rotate(-90)`);
-  }
-
-  prepCanvas (chartShape) {
-    // We only want to set canvas width / height ATTRIBUTES here, because doing
-    // so clears the canvas. In another context (e.g. zooming / panning), we
-    // only want to set transform STYLES to scale what we previously drew here)
-    this.d3el.select('.gantt-canvas')
-      .attr('width', chartShape.bins)
-      .attr('height', chartShape.fullHeight)
-      .style('transform-origin', null)
-      .style('transform', null);
+      .text('Location');
   }
 
   drawBars (chartShape) {
@@ -608,7 +260,7 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
     const totalUtilization = this.getNamedResource('totalUtilization');
     const selectionUtilization = this.getNamedResource('selectionUtilization');
 
-    const canvas = this.d3el.select('.gantt-canvas');
+    const canvas = this.d3el.select('canvas');
     const ctx = canvas.node().getContext('2d');
 
     const bandwidth = this.yScale.bandwidth();
@@ -646,7 +298,7 @@ class GanttView extends LinkedMixin( // Ensures that this.linkedState is updated
     }
     const theme = globalThis.controller.getNamedResource('theme').cssVariables;
 
-    const canvas = this.d3el.select('.gantt-canvas');
+    const canvas = this.d3el.select('canvas');
     const ctx = canvas.node().getContext('2d');
 
     const bandwidth = this.yScale.bandwidth();
