+++ static/views/ZoomableTimelineView/style.less
@@ -1,4 +1,4 @@
-.GanttView {
+.ZoomableTimelineView {
   .eventCapturer {
     cursor: grab;
     touch-action: none; /* Don't zoom the browser when the user pinches */
@@ -14,10 +14,10 @@
     stroke-weight: 1px;
     stroke: var(--border-color);
   }
-  .yScroller {
+  foreignObject {
     /* overflow-y set by javascript */
     overflow-x: hidden;
-    .gantt-canvas {
+    canvas {
       position: absolute;
     }
   }
